+++ src/menus.ts
@@ -1,20 +1,21 @@
-import GENE_CONVERT_MODULE from './components/pages/apps/genes/gene-convert/manifest.json'
-import GSEA_PLOT_MODULE from './components/pages/apps/genes/gsea-plot/manifest.json'
-import GSEA_WEB_MODULE from './components/pages/apps/genes/gsea-web/manifest.json'
-import MOTIFS_MODULE from './components/pages/apps/genes/motifs/manifest.json'
-import PATHWAY_MODULE from './components/pages/apps/genes/pathway/manifest.json'
-import ANNOTATE_MODULE from './components/pages/apps/genomic/annotate/manifest.json'
-import DNA_MODULE from './components/pages/apps/genomic/dna/manifest.json'
-import OVERLAP_MODULE from './components/pages/apps/genomic/overlap/manifest.json'
-import REV_COMP_MODULE from './components/pages/apps/genomic/rev-comp/manifest.json'
-import SEQBROWSER_MODULE from './components/pages/apps/genomic/seq-browser/manifest.json'
-import HUBS_MODULE from './components/pages/apps/hubs/manifest.json'
-import MATCALC_MODULE from './components/pages/apps/matcalc/manifest.json'
-import SINGLE_CELL_MODULE from './components/pages/apps/ngs/single-cell/manifest.json'
-import VENN_MODULE from './components/pages/apps/venn/manifest.json'
-import LOLLIPOP_MODULE from './components/pages/apps/wgs/lollipop/manifest.json'
-import ONCOPLOT_MODULE from './components/pages/apps/wgs/oncoplot/manifest.json'
-import VARIANTS_MODULE from './components/pages/apps/wgs/variants/manifest.json'
+import GENE_CONVERT_APP from './components/pages/apps/genes/gene-convert/manifest.json'
+import GSEA_PLOT_APP from './components/pages/apps/genes/gsea-plot/manifest.json'
+import GSEA_WEB_APP from './components/pages/apps/genes/gsea-web/manifest.json'
+import MOTIFS_APP from './components/pages/apps/genes/motifs/manifest.json'
+import PATHWAY_APP from './components/pages/apps/genes/pathway/manifest.json'
+import ANNOTATE_APP from './components/pages/apps/genomic/annotate/manifest.json'
+import DNA_APP from './components/pages/apps/genomic/dna/manifest.json'
+import OVERLAP_APP from './components/pages/apps/genomic/overlap/manifest.json'
+import REV_COMP_APP from './components/pages/apps/genomic/rev-comp/manifest.json'
+import SEQBROWSER_APP from './components/pages/apps/genomic/seq-browser/manifest.json'
+import HUBS_APP from './components/pages/apps/hubs/manifest.json'
+import MATCALC_APP from './components/pages/apps/matcalc/manifest.json'
+import SINGLE_CELL_APP from './components/pages/apps/ngs/single-cell/manifest.json'
+import SANKEY_APP from './components/pages/apps/sankey/manifest.json'
+import VENN_APP from './components/pages/apps/venn/manifest.json'
+import LOLLIPOP_APP from './components/pages/apps/wgs/lollipop/manifest.json'
+import ONCOPLOT_APP from './components/pages/apps/wgs/oncoplot/manifest.json'
+import VARIANTS_APP from './components/pages/apps/wgs/variants/manifest.json'
 
 export interface IAppHeaderLink {
   name: string
@@ -35,7 +36,7 @@ export const HEADER_LINKS: IHeaderLink[] = [
     name: 'Plot',
     apps: [
       {
-        ...MATCALC_MODULE,
+        ...MATCALC_APP,
         slug: '/apps/matcalc',
         mode: 'prod',
       },
@@ -47,23 +48,28 @@ export const HEADER_LINKS: IHeaderLink[] = [
         color: '#6495ED',
       },
       {
-        ...ONCOPLOT_MODULE,
+        ...ONCOPLOT_APP,
         slug: '/apps/wgs/oncoplot',
         mode: 'prod',
       },
       {
-        ...LOLLIPOP_MODULE,
+        ...LOLLIPOP_APP,
         slug: '/apps/wgs/lollipop',
         mode: 'prod',
       },
       {
-        ...VENN_MODULE,
+        ...VENN_APP,
         slug: '/apps/venn',
         mode: 'prod',
       },
       {
-        ...SINGLE_CELL_MODULE,
+        ...SINGLE_CELL_APP,
         slug: '/apps/ngs/single-cell',
+        mode: 'prod',
+      },
+      {
+        ...SANKEY_APP,
+        slug: '/apps/sankey',
         mode: 'prod',
       },
     ],
@@ -73,27 +79,27 @@ export const HEADER_LINKS: IHeaderLink[] = [
     name: 'Genes',
     apps: [
       {
-        ...PATHWAY_MODULE,
+        ...PATHWAY_APP,
         slug: '/apps/genes/pathway',
         mode: 'prod',
       },
       {
-        ...MOTIFS_MODULE,
+        ...MOTIFS_APP,
         slug: '/apps/genes/motifs',
         mode: 'prod',
       },
       {
-        ...GENE_CONVERT_MODULE,
+        ...GENE_CONVERT_APP,
         slug: '/apps/genes/convert',
         mode: 'prod',
       },
       {
-        ...GSEA_WEB_MODULE,
+        ...GSEA_WEB_APP,
         slug: '/apps/genes/gsea-web',
         mode: 'dev',
       },
       {
-        ...GSEA_PLOT_MODULE,
+        ...GSEA_PLOT_APP,
         slug: '/apps/genes/gsea-plot',
         mode: 'prod',
       },
@@ -103,27 +109,27 @@ export const HEADER_LINKS: IHeaderLink[] = [
     name: 'Genomic',
     apps: [
       {
-        ...ANNOTATE_MODULE,
+        ...ANNOTATE_APP,
         slug: '/apps/genomic/annotate',
         mode: 'prod',
       },
       {
-        ...OVERLAP_MODULE,
+        ...OVERLAP_APP,
         slug: '/apps/genomic/overlap',
         mode: 'dev',
       },
       {
-        ...DNA_MODULE,
+        ...DNA_APP,
         slug: '/apps/genomic/dna',
         mode: 'prod',
       },
       {
-        ...REV_COMP_MODULE,
+        ...REV_COMP_APP,
         slug: '/apps/genomic/rev-comp',
         mode: 'prod',
       },
       {
-        ...HUBS_MODULE,
+        ...HUBS_APP,
         slug: '/apps/hubs',
         mode: 'prod',
       },
@@ -134,7 +140,7 @@ export const HEADER_LINKS: IHeaderLink[] = [
       //   mode: 'dev',
       // },
       {
-        ...SEQBROWSER_MODULE,
+        ...SEQBROWSER_APP,
         slug: '/apps/genomic/seq-browser',
         mode: 'prod',
       },
@@ -144,7 +150,7 @@ export const HEADER_LINKS: IHeaderLink[] = [
     name: 'WGS',
     apps: [
       {
-        ...VARIANTS_MODULE,
+        ...VARIANTS_APP,
         slug: '/apps/wgs/variants',
         mode: 'prod',
       },